+++ bpdm-common-test/src/main/kotlin/org/eclipse/tractusx/bpdm/test/testdata/pool/BusinessPartnerVerboseValues.kt
@@ -405,7 +405,7 @@ object BusinessPartnerVerboseValues {
             confidenceCriteria = confidenceCriteria1,
             isParticipantData = false,
             createdAt = Instant.now(),
-            updatedAt = Instant.now(),
+            updatedAt = Instant.now()
         ),
         legalAddress = addressPartnerInvariant1.copy(bpnLegalEntity = firstBpnL, addressType = AddressType.LegalAddress)
     )
@@ -421,7 +421,7 @@ object BusinessPartnerVerboseValues {
             confidenceCriteria = confidenceCriteria2,
             isParticipantData = false,
             createdAt = Instant.now(),
-            updatedAt = Instant.now(),
+            updatedAt = Instant.now()
         ),
         legalAddress = addressPartnerInvariant2.copy(bpnLegalEntity = secondBpnL, addressType = AddressType.LegalAddress)
     )
@@ -437,7 +437,7 @@ object BusinessPartnerVerboseValues {
             confidenceCriteria = confidenceCriteria3,
             isParticipantData = false,
             createdAt = Instant.now(),
-            updatedAt = Instant.now(),
+            updatedAt = Instant.now()
         ),
         legalAddress = addressPartnerInvariant3.copy(bpnLegalEntity = thirdBpnl, addressType = AddressType.LegalAddress)
     )